Move most runtime files up one level
This avoids the `oak_runtime/runtime` stutter, and I think makes more
sense anyways.
