+++ oak/server/rust/oak_runtime/src/config.rs
@@ -16,7 +16,7 @@
 
 //! Functionality covering configuration of a Runtime instance.
 
-use crate::{runtime::RuntimeProxy, Runtime, RuntimeConfiguration};
+use crate::{Runtime, RuntimeConfiguration, RuntimeProxy};
 use log::{error, info};
 use oak_abi::OakStatus;
 use std::sync::Arc;
