+++ oak/server/rust/oak_runtime/src/introspection_events.rs
@@ -16,7 +16,7 @@
 
 use crate::{
     proto::oak::introspection_events::{event::EventDetails, Event},
-    runtime::Runtime,
+    Runtime,
 };
 use log::debug;
 use prost_types::Timestamp;
